Move "SpanExtensions" to LoRaWan project (for wider reuse)
