+++ AssemblyInfo.cs
@@ -6,6 +6,7 @@ using System.Runtime.CompilerServices;
 [assembly: System.CLSCompliant(false)]
 
 [assembly: InternalsVisibleTo("LoRaTools")]
+[assembly: InternalsVisibleTo("LoRaWan.NetworkServer")]
 [assembly: InternalsVisibleTo("LoRaWan.Tests.Unit")]
 [assembly: InternalsVisibleTo("LoRaWan.Tests.Common")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
